Merge Unbox() and UnboxOrThrow()
The Unbox() function was introduced before Swift 2 and the do/try/catch
model, so the UnboxOrThrow() was an easy way to ease in that pattern without
breaking existing usage. However, since the same behavior as the current
version of Unbox() has (it returns nil on failure) can be achieved with
try? UnboxOrThrow() there’s no need to keep maintaining two separate
functions. So, with this change, UnboxOrThrow() is renamed to Unbox() and
the old Unbox() function is removed.