Automatically convert strings into numbers
This change makes it so that if a number is being unboxed, but a string
was found, an attempt will be made to transform the string into a number
of the given type. That way, no manual transformation code needs to be
written, and the same simple unbox() API can be used.