Enable init(unboxer:) + init(unboxer:context:) to throw
This change changes the initializer signature for `Unboxable` and
`UnboxableWithContext` to `throw`. This enables a more quick exit out
of initializers whenever an invalid value was found, making Unbox faster.
This is accomplished by using try when unboxing a required value.

It also has the added benefit of enabling API users to manually exit out
of an initializer by throwing, instead of having to use the failing APIs
on `Unboxer`.

The only downside of this approach is that the error reporting goes back
to only returning a single error (as we’ll exit early, instead of gathering
all errors), however, for the sake of API streamlining + performance, this
is a tradeoff that is worth making.