Unboxer: Make initializers public
Now that Unboxer is stateless, we can enable the API user to construct
instances of it. Before, there was a risk that state would be shared
if Unboxer wasn’t initialized within the framework, but that’s no longer
the case.

This way of creating Unboxers is now an alternative to using the custom
unboxing API.