Unbox 3: Add migration script
Being a major release, Unbox 3.0 breaks the 2.x API. To make that impact
as small as possible on developers using Unbox, a migration script is
provided, that can super easily migrate code using Unbox 2.x to Unbox 3.0.

Instructions on how to run it will be added to the README in an upcoming
commit.