Deprecate Unbox
This change marks Unbox as deprecated, in favor of Swift’s built-in
`Codable` API and the Codextended project. Unbox will be kept online
for as long as possible (although soon it will be made read-only using
GitHub’s archive feature).

Added to this change is a migration guide which aims to make it easier
for developers using Unbox to start migrating to `Codable`.